Simplify 402 driver (#89)
* Split motor.hpp and motor.cpp into different files


* Fix missing symbol error


---------
