+++ canopen_402_driver/CMakeLists.txt
@@ -32,6 +32,9 @@ set(dependencies
 
 add_library(lely_motion_controller_bridge
   src/motor.cpp
+  src/command.cpp
+  src/state.cpp
+  src/default_homing_mode.cpp
   src/lely_motion_controller_bridge.cpp
 )
 target_compile_features(lely_motion_controller_bridge PUBLIC c_std_99 cxx_std_17)  # Require C99 and C++17
