+++ canopen_402_driver/include/canopen_402_driver/motor.hpp
@@ -12,240 +12,13 @@
 #include <thread>
 #include "rclcpp/rclcpp.hpp"
 
+#include "canopen_402_driver/default_homing_mode.hpp"
 #include "canopen_402_driver/lely_motion_controller_bridge.hpp"
-using namespace ros2_canopen;
+#include "canopen_402_driver/mode_forward_helper.hpp"
+#include "canopen_402_driver/profiled_position_mode.hpp"
+
 namespace ros2_canopen
 {
-class State402
-{
-public:
-  enum StatusWord
-  {
-    SW_Ready_To_Switch_On = 0,
-    SW_Switched_On = 1,
-    SW_Operation_enabled = 2,
-    SW_Fault = 3,
-    SW_Voltage_enabled = 4,
-    SW_Quick_stop = 5,
-    SW_Switch_on_disabled = 6,
-    SW_Warning = 7,
-    SW_Manufacturer_specific0 = 8,
-    SW_Remote = 9,
-    SW_Target_reached = 10,
-    SW_Internal_limit = 11,
-    SW_Operation_mode_specific0 = 12,
-    SW_Operation_mode_specific1 = 13,
-    SW_Manufacturer_specific1 = 14,
-    SW_Manufacturer_specific2 = 15
-  };
-  enum InternalState
-  {
-    Unknown = 0,
-    Start = 0,
-    Not_Ready_To_Switch_On = 1,
-    Switch_On_Disabled = 2,
-    Ready_To_Switch_On = 3,
-    Switched_On = 4,
-    Operation_Enable = 5,
-    Quick_Stop_Active = 6,
-    Fault_Reaction_Active = 7,
-    Fault = 8,
-  };
-  InternalState getState();
-  InternalState read(uint16_t sw);
-  bool waitForNewState(
-    const std::chrono::steady_clock::time_point & abstime, InternalState & state);
-  State402() : state_(Unknown) {}
-
-private:
-  std::condition_variable cond_;
-  std::mutex mutex_;
-  InternalState state_;
-};
-
-class Command402
-{
-  struct Op
-  {
-    uint16_t to_set_;
-    uint16_t to_reset_;
-    Op(uint16_t to_set, uint16_t to_reset) : to_set_(to_set), to_reset_(to_reset) {}
-    void operator()(uint16_t & val) const { val = (val & ~to_reset_) | to_set_; }
-  };
-  class TransitionTable
-  {
-    boost::container::flat_map<std::pair<State402::InternalState, State402::InternalState>, Op>
-      transitions_;
-    void add(const State402::InternalState & from, const State402::InternalState & to, Op op)
-    {
-      transitions_.insert(std::make_pair(std::make_pair(from, to), op));
-    }
-
-  public:
-    TransitionTable();
-    const Op & get(const State402::InternalState & from, const State402::InternalState & to) const
-    {
-      return transitions_.at(std::make_pair(from, to));
-    }
-  };
-  static const TransitionTable transitions_;
-  static State402::InternalState nextStateForEnabling(State402::InternalState state);
-  Command402();
-
-public:
-  enum ControlWord
-  {
-    CW_Switch_On = 0,
-    CW_Enable_Voltage = 1,
-    CW_Quick_Stop = 2,
-    CW_Enable_Operation = 3,
-    CW_Operation_mode_specific0 = 4,
-    CW_Operation_mode_specific1 = 5,
-    CW_Operation_mode_specific2 = 6,
-    CW_Fault_Reset = 7,
-    CW_Halt = 8,
-    CW_Operation_mode_specific3 = 9,
-    // CW_Reserved1=10,
-    CW_Manufacturer_specific0 = 11,
-    CW_Manufacturer_specific1 = 12,
-    CW_Manufacturer_specific2 = 13,
-    CW_Manufacturer_specific3 = 14,
-    CW_Manufacturer_specific4 = 15,
-  };
-  static bool setTransition(
-    uint16_t & cw, const State402::InternalState & from, const State402::InternalState & to,
-    State402::InternalState * next);
-};
-
-template <uint16_t MASK>
-class WordAccessor
-{
-  uint16_t & word_;
-
-public:
-  WordAccessor(uint16_t & word) : word_(word) {}
-  bool set(uint8_t bit)
-  {
-    uint16_t val = MASK & (1 << bit);
-    word_ |= val;
-    return val;
-  }
-  bool reset(uint8_t bit)
-  {
-    uint16_t val = MASK & (1 << bit);
-    word_ &= ~val;
-    return val;
-  }
-  bool get(uint8_t bit) const { return word_ & (1 << bit); }
-  uint16_t get() const { return word_ & MASK; }
-  WordAccessor & operator=(const uint16_t & val)
-  {
-    word_ = (word_ & ~MASK) | (val & MASK);
-    return *this;
-  }
-};
-
-class Mode
-{
-public:
-  const uint16_t mode_id_;
-  Mode(uint16_t id) : mode_id_(id) {}
-  typedef WordAccessor<
-    (1 << Command402::CW_Operation_mode_specific0) |
-    (1 << Command402::CW_Operation_mode_specific1) |
-    (1 << Command402::CW_Operation_mode_specific2) | (1 << Command402::CW_Operation_mode_specific3)>
-    OpModeAccesser;
-  virtual bool start() = 0;
-  virtual bool read(const uint16_t & sw) = 0;
-  virtual bool write(OpModeAccesser & cw) = 0;
-  virtual bool setTarget(const double & val) { return false; }
-  virtual ~Mode() {}
-};
-typedef std::shared_ptr<Mode> ModeSharedPtr;
-
-template <typename T>
-class ModeTargetHelper : public Mode
-{
-  T target_;
-  std::atomic<bool> has_target_;
-
-public:
-  ModeTargetHelper(uint16_t mode) : Mode(mode) {}
-  bool hasTarget() { return has_target_; }
-  T getTarget() { return target_; }
-  virtual bool setTarget(const double & val)
-  {
-    if (std::isnan(val))
-    {
-      std::cout << "canopen_402 target command is not a number" << std::endl;
-      return false;
-    }
-
-    using boost::numeric_cast;
-    using boost::numeric::negative_overflow;
-    using boost::numeric::positive_overflow;
-
-    try
-    {
-      target_ = numeric_cast<T>(val);
-    }
-    catch (negative_overflow &)
-    {
-      std::cout << "canopen_402 Command " << val
-                << " does not fit into target, clamping to min limit" << std::endl;
-      target_ = std::numeric_limits<T>::min();
-    }
-    catch (positive_overflow &)
-    {
-      std::cout << "canopen_402 Command " << val
-                << " does not fit into target, clamping to max limit" << std::endl;
-      target_ = std::numeric_limits<T>::max();
-    }
-    catch (...)
-    {
-      std::cout << "canopen_402 Was not able to cast command " << val << std::endl;
-      return false;
-    }
-
-    has_target_ = true;
-    return true;
-  }
-  virtual bool start()
-  {
-    has_target_ = false;
-    return true;
-  }
-};
-
-template <uint16_t ID, typename TYPE, CODataTypes TPY, uint16_t OBJ, uint8_t SUB, uint16_t CW_MASK>
-class ModeForwardHelper : public ModeTargetHelper<TYPE>
-{
-  std::shared_ptr<LelyMotionControllerBridge> driver;
-  std::shared_ptr<RemoteObject> obj;
-
-public:
-  ModeForwardHelper(std::shared_ptr<LelyMotionControllerBridge> driver) : ModeTargetHelper<TYPE>(ID)
-  {
-    this->obj = driver->create_remote_obj(OBJ, SUB, TPY);
-    this->driver = driver;
-  }
-  virtual bool read(const uint16_t & sw) { return true; }
-  virtual bool write(Mode::OpModeAccesser & cw)
-  {
-    if (this->hasTarget())
-    {
-      cw = cw.get() | CW_MASK;
-
-      driver->set_remote_obj(obj, this->getTarget());
-      return true;
-    }
-    else
-    {
-      cw = cw.get() & ~CW_MASK;
-      return false;
-    }
-  }
-};
 
 typedef ModeForwardHelper<
   MotorBase::Profiled_Velocity, int32_t, CODataTypes::COData32, 0x60FF, 0, 0>
@@ -270,131 +43,6 @@ typedef ModeForwardHelper<
   MotorBase::Interpolated_Position, int32_t, CODataTypes::COData32, 0x60C1, 0x01,
   (1 << Command402::CW_Operation_mode_specific0)>
   InterpolatedPositionMode;
-
-class ProfiledPositionMode : public ModeTargetHelper<int32_t>
-{
-  const uint16_t index = 0x607A;
-  std::shared_ptr<LelyMotionControllerBridge> driver;
-  std::shared_ptr<RemoteObject> obj;
-
-  double last_target_;
-  uint16_t sw_;
-
-public:
-  enum SW_masks
-  {
-    MASK_Reached = (1 << State402::SW_Target_reached),
-    MASK_Acknowledged = (1 << State402::SW_Operation_mode_specific0),
-    MASK_Error = (1 << State402::SW_Operation_mode_specific1),
-  };
-  enum CW_bits
-  {
-    CW_NewPoint = Command402::CW_Operation_mode_specific0,
-    CW_Immediate = Command402::CW_Operation_mode_specific1,
-    CW_Blending = Command402::CW_Operation_mode_specific3,
-  };
-  ProfiledPositionMode(std::shared_ptr<LelyMotionControllerBridge> driver)
-  : ModeTargetHelper(MotorBase::Profiled_Position)
-  {
-    this->driver = driver;
-    obj = driver->create_remote_obj(index, 0U, CODataTypes::COData32);
-  }
-
-  virtual bool start()
-  {
-    sw_ = 0;
-    last_target_ = std::numeric_limits<double>::quiet_NaN();
-    return ModeTargetHelper::start();
-  }
-  virtual bool read(const uint16_t & sw)
-  {
-    sw_ = sw;
-    return (sw & MASK_Error) == 0;
-  }
-  virtual bool write(OpModeAccesser & cw)
-  {
-    cw.set(CW_Immediate);
-    if (hasTarget())
-    {
-      int32_t target = getTarget();
-      if ((sw_ & MASK_Acknowledged) == 0 && target != last_target_)
-      {
-        if (cw.get(CW_NewPoint))
-        {
-          cw.reset(CW_NewPoint);  // reset if needed
-        }
-        else
-        {
-          driver->set_remote_obj(obj, target);
-          cw.set(CW_NewPoint);
-          last_target_ = target;
-        }
-      }
-      else if (sw_ & MASK_Acknowledged)
-      {
-        cw.reset(CW_NewPoint);
-      }
-      return true;
-    }
-    return false;
-  }
-};
-
-class HomingMode : public Mode
-{
-protected:
-  enum SW_bits
-  {
-    SW_Attained = State402::SW_Operation_mode_specific0,
-    SW_Error = State402::SW_Operation_mode_specific1,
-  };
-  enum CW_bits
-  {
-    CW_StartHoming = Command402::CW_Operation_mode_specific0,
-  };
-
-public:
-  HomingMode() : Mode(MotorBase::Homing) {}
-  virtual bool executeHoming() = 0;
-};
-
-class DefaultHomingMode : public HomingMode
-{
-  const uint16_t index = 0x6098;
-  std::shared_ptr<LelyMotionControllerBridge> driver;
-  std::shared_ptr<RemoteObject> obj;
-
-  std::atomic<bool> execute_;
-
-  std::mutex mutex_;
-  std::condition_variable cond_;
-  uint16_t status_;
-
-  enum SW_masks
-  {
-    MASK_Reached = (1 << State402::SW_Target_reached),
-    MASK_Attained = (1 << SW_Attained),
-    MASK_Error = (1 << SW_Error),
-  };
-  bool error(const std::string & msg)
-  {
-    execute_ = false;
-    std::cout << msg << std::endl;
-    return false;
-  }
-
-public:
-  DefaultHomingMode(std::shared_ptr<LelyMotionControllerBridge> driver)
-  {
-    this->driver = driver;
-    obj = driver->create_remote_obj(index, 0U, CODataTypes::COData32);
-  }
-  virtual bool start();
-  virtual bool read(const uint16_t & sw);
-  virtual bool write(OpModeAccesser & cw);
-
-  virtual bool executeHoming();
-};
 
 class Motor402 : public MotorBase
 {
