+++ canopen_402_driver/src/motor.cpp
@@ -1,294 +1,6 @@
 #include "canopen_402_driver/motor.hpp"
 using namespace ros2_canopen;
 
-State402::InternalState State402::getState()
-{
-  std::scoped_lock lock(mutex_);
-  return state_;
-}
-
-State402::InternalState State402::read(uint16_t sw)
-{
-  static const uint16_t r = (1 << SW_Ready_To_Switch_On);
-  static const uint16_t s = (1 << SW_Switched_On);
-  static const uint16_t o = (1 << SW_Operation_enabled);
-  static const uint16_t f = (1 << SW_Fault);
-  static const uint16_t q = (1 << SW_Quick_stop);
-  static const uint16_t d = (1 << SW_Switch_on_disabled);
-
-  InternalState new_state = Unknown;
-
-  uint16_t state = sw & (d | q | f | o | s | r);
-  switch (state)
-  {
-    //   ( d | q | f | o | s | r ):
-    case (0 | 0 | 0 | 0 | 0 | 0):
-    case (0 | q | 0 | 0 | 0 | 0):
-      // std::cout << "Not_Ready_To_Switch_On" << std::endl;
-      new_state = Not_Ready_To_Switch_On;
-      break;
-
-    case (d | 0 | 0 | 0 | 0 | 0):
-    case (d | q | 0 | 0 | 0 | 0):
-      // std::cout << "Switch_On_Disabled" << std::endl;
-      new_state = Switch_On_Disabled;
-      break;
-
-    case (0 | q | 0 | 0 | 0 | r):
-      // std::cout << "Ready_To_Switch_On" << std::endl;
-      new_state = Ready_To_Switch_On;
-      break;
-
-    case (0 | q | 0 | 0 | s | r):
-      new_state = Switched_On;
-      break;
-
-    case (0 | q | 0 | o | s | r):
-      new_state = Operation_Enable;
-      break;
-
-    case (0 | 0 | 0 | o | s | r):
-      new_state = Quick_Stop_Active;
-      break;
-
-    case (0 | 0 | f | o | s | r):
-    case (0 | q | f | o | s | r):
-      new_state = Fault_Reaction_Active;
-      break;
-
-    case (0 | 0 | f | 0 | 0 | 0):
-    case (0 | q | f | 0 | 0 | 0):
-      new_state = Fault;
-      break;
-
-    default:
-      /// @todo Throw error here.
-      break;
-  }
-  std::scoped_lock lock(mutex_);
-  if (new_state != state_)
-  {
-    state_ = new_state;
-    cond_.notify_all();
-  }
-  return state_;
-}
-bool State402::waitForNewState(
-  const std::chrono::steady_clock::time_point & abstime, State402::InternalState & state)
-{
-  std::unique_lock lock(mutex_);
-  while (state_ == state && cond_.wait_until(lock, abstime) == std::cv_status::no_timeout)
-  {
-  }
-  bool res = state != state_;
-  state = state_;
-  return res;
-}
-
-const Command402::TransitionTable Command402::transitions_;
-
-Command402::TransitionTable::TransitionTable()
-{
-  typedef State402 s;
-
-  transitions_.reserve(32);
-
-  Op disable_voltage(0, (1 << CW_Fault_Reset) | (1 << CW_Enable_Voltage));
-  /* 7*/ add(s::Ready_To_Switch_On, s::Switch_On_Disabled, disable_voltage);
-  /* 9*/ add(s::Operation_Enable, s::Switch_On_Disabled, disable_voltage);
-  /*10*/ add(s::Switched_On, s::Switch_On_Disabled, disable_voltage);
-  /*12*/ add(s::Quick_Stop_Active, s::Switch_On_Disabled, disable_voltage);
-
-  Op automatic(0, 0);
-  /* 0*/ add(s::Start, s::Not_Ready_To_Switch_On, automatic);
-  /* 1*/ add(s::Not_Ready_To_Switch_On, s::Switch_On_Disabled, automatic);
-  /*14*/ add(s::Fault_Reaction_Active, s::Fault, automatic);
-
-  Op shutdown(
-    (1 << CW_Quick_Stop) | (1 << CW_Enable_Voltage), (1 << CW_Fault_Reset) | (1 << CW_Switch_On));
-  /* 2*/ add(s::Switch_On_Disabled, s::Ready_To_Switch_On, shutdown);
-  /* 6*/ add(s::Switched_On, s::Ready_To_Switch_On, shutdown);
-  /* 8*/ add(s::Operation_Enable, s::Ready_To_Switch_On, shutdown);
-
-  Op switch_on(
-    (1 << CW_Quick_Stop) | (1 << CW_Enable_Voltage) | (1 << CW_Switch_On),
-    (1 << CW_Fault_Reset) | (1 << CW_Enable_Operation));
-  /* 3*/ add(s::Ready_To_Switch_On, s::Switched_On, switch_on);
-  /* 5*/ add(s::Operation_Enable, s::Switched_On, switch_on);
-
-  Op enable_operation(
-    (1 << CW_Quick_Stop) | (1 << CW_Enable_Voltage) | (1 << CW_Switch_On) |
-      (1 << CW_Enable_Operation),
-    (1 << CW_Fault_Reset));
-  /* 4*/ add(s::Switched_On, s::Operation_Enable, enable_operation);
-  /*16*/ add(s::Quick_Stop_Active, s::Operation_Enable, enable_operation);
-
-  Op quickstop((1 << CW_Enable_Voltage), (1 << CW_Fault_Reset) | (1 << CW_Quick_Stop));
-  /* 7*/ add(
-    s::Ready_To_Switch_On, s::Quick_Stop_Active, quickstop);    // transit to Switch_On_Disabled
-  /*10*/ add(s::Switched_On, s::Quick_Stop_Active, quickstop);  // transit to Switch_On_Disabled
-  /*11*/ add(s::Operation_Enable, s::Quick_Stop_Active, quickstop);
-
-  // fault reset
-  /*15*/ add(s::Fault, s::Switch_On_Disabled, Op((1 << CW_Fault_Reset), 0));
-}
-State402::InternalState Command402::nextStateForEnabling(State402::InternalState state)
-{
-  switch (state)
-  {
-    case State402::Start:
-      return State402::Not_Ready_To_Switch_On;
-
-    case State402::Fault:
-    case State402::Not_Ready_To_Switch_On:
-      return State402::Switch_On_Disabled;
-
-    case State402::Switch_On_Disabled:
-      return State402::Ready_To_Switch_On;
-
-    case State402::Ready_To_Switch_On:
-      return State402::Switched_On;
-
-    case State402::Switched_On:
-    case State402::Quick_Stop_Active:
-    case State402::Operation_Enable:
-      return State402::Operation_Enable;
-
-    case State402::Fault_Reaction_Active:
-      return State402::Fault;
-  }
-  throw std::out_of_range("state value is illegal");
-}
-
-bool Command402::setTransition(
-  uint16_t & cw, const State402::InternalState & from, const State402::InternalState & to,
-  State402::InternalState * next)
-{
-  try
-  {
-    if (from != to)
-    {
-      State402::InternalState hop = to;
-      if (next)
-      {
-        if (to == State402::Operation_Enable) hop = nextStateForEnabling(from);
-        *next = hop;
-      }
-      transitions_.get(from, hop)(cw);
-    }
-    return true;
-  }
-  catch (...)
-  {
-    /// @todo Print error here.
-  }
-  return false;
-}
-
-template <uint16_t mask, uint16_t not_equal>
-struct masked_status_not_equal
-{
-  uint16_t & status_;
-  masked_status_not_equal(uint16_t & status) : status_(status) {}
-  bool operator()() const { return (status_ & mask) != not_equal; }
-};
-bool DefaultHomingMode::start()
-{
-  execute_ = false;
-  return read(0);
-}
-bool DefaultHomingMode::read(const uint16_t & sw)
-{
-  std::scoped_lock lock(mutex_);
-  uint16_t old = status_;
-  status_ = sw & (MASK_Reached | MASK_Attained | MASK_Error);
-  if (old != status_)
-  {
-    cond_.notify_all();
-  }
-  return true;
-}
-bool DefaultHomingMode::write(Mode::OpModeAccesser & cw)
-{
-  cw = 0;
-  if (execute_)
-  {
-    cw.set(CW_StartHoming);
-    return true;
-  }
-  return false;
-}
-
-bool DefaultHomingMode::executeHoming()
-{
-  int hmode = driver->get_remote_obj<int8_t>(obj);
-  if (hmode == 0)
-  {
-    return true;
-  }
-  /// @ get abs time from canopen_master
-  std::chrono::steady_clock::time_point prepare_time =
-    std::chrono::steady_clock::now() + std::chrono::seconds(1);
-  // ensure homing is not running
-  std::unique_lock lock(mutex_);
-  if (!cond_.wait_until(
-        lock, prepare_time, masked_status_not_equal<MASK_Error | MASK_Reached, 0>(status_)))
-  {
-    return error("could not prepare homing");
-  }
-  if (status_ & MASK_Error)
-  {
-    return error("homing error before start");
-  }
-
-  execute_ = true;
-
-  // ensure start
-  if (!cond_.wait_until(
-        lock, prepare_time,
-        masked_status_not_equal<MASK_Error | MASK_Attained | MASK_Reached, MASK_Reached>(status_)))
-  {
-    return error("homing did not start");
-  }
-  if (status_ & MASK_Error)
-  {
-    return error("homing error at start");
-  }
-
-  std::chrono::steady_clock::time_point finish_time =
-    std::chrono::steady_clock::now() + std::chrono::seconds(10);  //
-
-  // wait for attained
-  if (!cond_.wait_until(
-        lock, finish_time, masked_status_not_equal<MASK_Error | MASK_Attained, 0>(status_)))
-  {
-    return error("homing not attained");
-  }
-  if (status_ & MASK_Error)
-  {
-    return error("homing error during process");
-  }
-
-  // wait for motion stop
-  if (!cond_.wait_until(
-        lock, finish_time, masked_status_not_equal<MASK_Error | MASK_Reached, 0>(status_)))
-  {
-    return error("homing did not stop");
-  }
-  if (status_ & MASK_Error)
-  {
-    return error("homing error during stop");
-  }
-
-  if ((status_ & MASK_Reached) && (status_ & MASK_Attained))
-  {
-    execute_ = false;
-    return true;
-  }
-
-  return error("something went wrong while homing");
-}
-
 bool Motor402::setTarget(double val)
 {
   if (state_handler_.getState() == State402::Operation_Enable)
